requirements.txt
ansible==2.10.7
ansible-base==2.10.16
awscli==1.18.214
boto3==1.16.54
botocore==1.19.54
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==3.3.2
docutils==0.15.2
idna==2.8
Jinja2==3.0.3
jmespath==0.10.0
MarkupSafe==2.0.1
packaging==21.3
psycopg2==2.9.2
pyasn1==0.4.8
pycparser==2.21
pyOpenSSL==20.0.1
pyparsing==3.0.6
python-dateutil==2.8.2
PyYAML==5.3.1
requests==2.22.0
rsa==4.5
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
s3transfer==0.3.7
six==1.16.0
urllib3==1.25.11
